feat: Add SettingsWidget component for user profile management
- Implemented SettingsWidget to allow users to edit their profile information including name, phone, and address.
- Integrated phone and address validation with user feedback.
- Added SettingsWidget to ProtectedLayout for accessibility across authenticated pages.
- Updated Sidebar to replace logo with an image.
- Refactored TopNavigation to accept a title prop and removed inline header code from various pages (AddExpensePage, DashboardPage, InsightsPage, ReportPage).
- Adjusted CSS styles for the new SettingsWidget and modified existing styles for better layout on mobile.
