+++ app/src/layouts/ProtectedLayout.jsx
@@ -1,18 +1,13 @@
 import Sidebar from '../components/Sidebar';
 import ExpenseChatWidget from '../components/ExpenseChatWidget';
+import SettingsWidget from '../components/SettingsWidget';
 
 /**
  * ProtectedLayout wraps all authenticated pages with:
  * - Desktop: collapsible left sidebar (sidebar-proto)
  * - Mobile: fixed bottom nav bar (sidebar-proto-mobile-nav)
  * - ExpenseChatWidget floating in bottom-right
- *
- * Pages inside this layout no longer need their own dashboard-header /
- * TopNavigation / Logout button — those are handled here.
- * Exception: DashboardPage, ReportPage, AddExpensePage, InsightsPage
- * currently render their own header; this layout sits around them without
- * conflict because the sidebar and bottom-nav are independent of the
- * per-page header area.
+ * - SettingsWidget floating near chat widget
  */
 export default function ProtectedLayout({ children }) {
   return (
@@ -22,6 +17,7 @@ export default function ProtectedLayout({ children }) {
         {children}
       </div>
       <ExpenseChatWidget />
+      <SettingsWidget />
     </div>
   );
 }
