+++ app/src/styles/components.css
@@ -268,10 +268,7 @@
 }
 
 /* --- Expense Chat Widget --- */
-.expense-chat-widget {
-  position: fixed;
-  right: 20px;
-  bottom: 20px;
+`$120px;
   z-index: 60;
   display: grid;
   justify-items: end;
@@ -419,3 +416,40 @@
   font-size: 13px;
   color: var(--muted);
 }
+
+/* --- Settings Widget --- */
+.settings-widget-container {
+  position: fixed;
+  right: 20px;
+  bottom: 90px;
+  z-index: 60;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  gap: 12px;
+}
+
+.settings-widget-launcher {
+  width: 58px;
+  height: 58px;
+  border-radius: 999px;
+  padding: 0;
+  display: inline-grid;
+  place-items: center;
+  box-shadow: 0 16px 34px rgba(13, 31, 62, 0.16);
+  background: var(--surface);
+  color: var(--brand);
+  border: 1px solid var(--brand-soft-border);
+  cursor: pointer;
+}
+
+/* Base adjustment for mobile bottom-nav stack */
+@media (max-width: 768px) {
+  .expense-chat-widget {
+    bottom: 80px;
+  }
+  .settings-widget-container {
+    bottom: 150px;
+  }
+}
+
